requirements.txt
pytest==7.4.2
python-dotenv==1.0.0
allure-pytest==2.13.2
curlify==2.2.1
selene==2.0.0rc4